Introduce a new utils.Graph class
This new class intends to replace the usage of vascpy.PointVasculature
instances in the code and get rid of the `GRAPH_HELPER` class whose
API was inconvenient, and its cache implementation dangerous.

An instance of `utils.Graph` is a `vascpy.PointVasculature` but with
additional methods and cached properties related to graph computation

`utils.Graph` constructor also sets its edge data so that there is
no need to call it explicitely like it was done for `vascpy.PointVasculature`.